restintern1/requirements.txt
asgiref==3.7.2
Django==4.1.13
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
Pillow==10.0.0
PyJWT==2.8.0
pymongo==3.12.1
pytz==2023.3
sqlparse==0.2.4
typing_extensions==4.7.1
tzdata==2023.3
